# cmd/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateEventDTO:
    type: object
  dto.EventResponseDTO:
    properties:
      created_at:
        type: string
      description:
        $ref: '#/definitions/entity.TranslatedText'
      end_date:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/entity.TranslatedText'
      main_photo_path:
        type: string
      photos:
        items:
          $ref: '#/definitions/dto.PhotoResponseDTO'
        type: array
      position:
        type: integer
      preview_photo_path:
        type: string
      start_date:
        type: string
      title:
        $ref: '#/definitions/entity.TranslatedText'
      updated_at:
        type: string
    type: object
...
  dto.PhotoResponseDTO:
    properties:
      id:
        type: integer
      path:
        type: string
      position:
        type: integer
    type: object
...
  dto.UpdateEventDTO:
    type: object
...
  entity.Photo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_main:
        type: boolean
      is_preview:
        type: boolean
      owner_id:
        type: integer
      owner_type:
        description: '"arts", "event"'
        type: string
      path:
        type: string
      position:
        type: integer
      updated_at:
        type: string
    type: object
...
  entity.TranslatedText:
    properties:
      en:
        example: Tung, tung, tung, tung, tung, tung, tung, tung, tung, Sahur.
        type: string
      es:
        example: Golybini Shpionini
        type: string
      ru:
        example: Тун, тун, тун, Сахур.
        type: string
    type: object
host: localhost:8010
info:
  contact: {}
  description: Backend для галереи anastasii_gofman
  title: Art Gallery API
  version: "1.0"
paths:
...
  /api/events:
    get:
      consumes:
      - application/json
      description: Получает все события
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/dto.EventResponseDTO'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all events
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: 'Создает новое событие без фотографий - принимает JSON. Поддерживаемые
        форматы дат: "2024-01-01T00:00:00Z", "01.01.2024", "01.01.2024 15:30"'
      parameters:
      - description: Event data with flexible date formats
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create event
      tags:
      - Events
  /api/events/{id}:
    get:
      consumes:
      - application/json
      description: Получает событие по ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event by ID
      tags:
      - Events
    patch:
      consumes:
      - application/json
      description: Обновляет поля события, которые ты передал, остальные не трогает,
        фото сюда не суй!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Partial update event
      tags:
      - Events
    put:
      consumes:
      - application/json
      description: Обновляет все поля события - все поля необязательные, но перезаписывает
        все поля, те которые не передал - обнуляются (кроме фото)
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Full update event
      tags:
      - Events
  /api/events/{id}/main_photo:
    get:
      consumes:
      - application/json
      description: Получает главную фотографию события
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Photo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get main photo
      tags:
      - Events
    post:
      consumes:
      - multipart/form-data
      description: Добавляет/обновляет главную фотографию к событию - передается фото
        в формате multipart/form-data в поле main_photo
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Main Photo
        in: formData
        name: main_photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add main photo to event
      tags:
      - Events
  /api/events/{id}/photos:
    patch:
      consumes:
      - multipart/form-data
      description: Обновляет фотографии события(имеющиеся удаляются) - передается
        []base64 в формате multipart/form-data в поле photos
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: Photos
        in: formData
        items:
          type: file
        name: photos
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch event photos
      tags:
      - Events
    post:
      consumes:
      - multipart/form-data
      description: Добавляет фотографии к событию(имеющиеся не трогаются) - передается
        []base64 в формате multipart/form-data в поле photos
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: Photos
        in: formData
        items:
          type: file
        name: photos
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add photos to event
      tags:
      - Events
  /api/events/with_photos:
    post:
      consumes:
      - multipart/form-data
      description: 'Создает событие и фотографии к нему - передается в формате multipart/form-data.
        Поддерживаемые форматы дат: "2024-01-01T00:00:00Z", "01.01.2024", "01.01.2024
        15:30"'
      parameters:
      - description: JSON payload for Event details with flexible date formats.
        in: formData
        name: data
        required: true
        type: string
        x-example: '{ "title": {"en": "Event Title", "ru": "Заголовок События", "es":
          "Título del Evento"}, "description": {"en": "Event description", "ru": "Описание
          события", "es": "Descripción del evento"}, "location": {"en": "Event location",
          "ru": "Место проведения", "es": "Ubicación del evento"}, "start_date": "01.01.2024",
          "end_date": "02.01.2024" }'
      - description: Main Photo
        in: formData
        name: main_photo
        type: file
      - description: Preview Photo
        in: formData
        name: preview_photo
        type: file
      - collectionFormat: csv
        description: Photos
        in: formData
        items:
          type: file
        name: photos
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EventResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create event with photos
      tags:
      - Events
swagger: "2.0"